chore: fix useEffect and move exported function
