+++ src/homepageExperience/components/steps/cli/ExecuteAggregateQuery.tsx
@@ -8,7 +8,7 @@ import {DEFAULT_BUCKET} from 'src/writeData/components/WriteDataDetailsContext'
 // Utils
 import {SafeBlankLink} from 'src/utils/SafeBlankLink'
 import {event} from 'src/cloud/utils/reporting'
-import {keyboardCopyTriggered} from 'src/utils/keyboardCopy'
+import {keyboardCopyTriggered} from 'src/utils/crossPlatform'
 
 const logCopyCodeSnippet = () => {
   event('firstMile.cliWizard.executeAggregateQuery.code.copied')
@@ -33,22 +33,21 @@ export const ExecuteAggregateQuery = (props: OwnProps) => {
   const codeSnippet = `influx query 'from(bucket:"${bucketName}") |> range(start:-10m) |> mean()' --raw`
 
   useEffect(() => {
+    const userSelection = () => {
+      return window.getSelection().toString()
+    }
+
+    const fireKeyboardCopyEvent = event => {
+      if (
+        keyboardCopyTriggered(event) &&
+        userSelection().includes('influx query')
+      ) {
+        logCopyCodeSnippet()
+      }
+    }
     document.addEventListener('keydown', fireKeyboardCopyEvent)
     return () => document.removeEventListener('keydown', fireKeyboardCopyEvent)
   }, [])
-
-  const userSelection = () => {
-    return window.getSelection().toString()
-  }
-
-  const fireKeyboardCopyEvent = event => {
-    if (
-      keyboardCopyTriggered(event) &&
-      userSelection().includes('influx query')
-    ) {
-      logCopyCodeSnippet()
-    }
-  }
 
   return (
     <>
