+++ src/homepageExperience/components/steps/cli/ExecuteQuery.tsx
@@ -10,7 +10,7 @@ import {DEFAULT_BUCKET} from 'src/writeData/components/WriteDataDetailsContext'
 // Utils
 import {event} from 'src/cloud/utils/reporting'
 import {SafeBlankLink} from 'src/utils/SafeBlankLink'
-import {keyboardCopyTriggered} from 'src/utils/keyboardCopy'
+import {keyboardCopyTriggered} from 'src/utils/crossPlatform'
 
 const logCopyCodeSnippet = () => {
   event('firstMile.cliWizard.executeQuery.code.copied')
@@ -30,22 +30,21 @@ export const ExecuteQuery = (props: OwnProps) => {
   |> filter(fn: (r) => r._measurement == “temperature”)`
 
   useEffect(() => {
+    const userSelection = () => {
+      return window.getSelection().toString()
+    }
+
+    const fireKeyboardCopyEvent = event => {
+      if (
+        keyboardCopyTriggered(event) &&
+        userSelection().includes('influx query')
+      ) {
+        logCopyCodeSnippet()
+      }
+    }
     document.addEventListener('keydown', fireKeyboardCopyEvent)
     return () => document.removeEventListener('keydown', fireKeyboardCopyEvent)
   }, [])
-
-  const userSelection = () => {
-    return window.getSelection().toString()
-  }
-
-  const fireKeyboardCopyEvent = event => {
-    if (
-      keyboardCopyTriggered(event) &&
-      userSelection().includes('influx query')
-    ) {
-      logCopyCodeSnippet()
-    }
-  }
 
   return (
     <>
