+++ src/homepageExperience/components/steps/cli/InitializeClient.tsx
@@ -31,7 +31,7 @@ import {WriteDataDetailsContext} from 'src/writeData/components/WriteDataDetails
 // Utils
 import {allAccessPermissions} from 'src/authorizations/utils/permissions'
 import {event} from 'src/cloud/utils/reporting'
-import {keyboardCopyTriggered} from 'src/utils/keyboardCopy'
+import {keyboardCopyTriggered} from 'src/utils/crossPlatform'
 
 // Types
 import {AppState, Authorization} from 'src/types'
@@ -115,6 +115,19 @@ export const InitializeClient: FC<OwnProps> = ({
   }, [currentAuth.token])
 
   useEffect(() => {
+    const userSelection = () => {
+      return window.getSelection().toString()
+    }
+
+    const fireKeyboardCopyEvent = event => {
+      if (
+        keyboardCopyTriggered(event) &&
+        (userSelection().includes('influx config create') ||
+          userSelection().includes('influx bucket create'))
+      ) {
+        logCopyCodeSnippet()
+      }
+    }
     document.addEventListener('keydown', fireKeyboardCopyEvent)
     return () => document.removeEventListener('keydown', fireKeyboardCopyEvent)
   }, [])
@@ -122,20 +135,6 @@ export const InitializeClient: FC<OwnProps> = ({
   // Events log handling
   const logCopyCodeSnippet = () => {
     event(`firstMile.${wizardEventName}.buckets.code.copied`)
-  }
-
-  const userSelection = () => {
-    return window.getSelection().toString()
-  }
-
-  const fireKeyboardCopyEvent = event => {
-    if (
-      keyboardCopyTriggered(event) &&
-      (userSelection().includes('influx config create') ||
-        userSelection().includes('influx bucket create'))
-    ) {
-      logCopyCodeSnippet()
-    }
   }
 
   return (
