+++ src/homepageExperience/components/steps/cli/InstallDependencies.tsx
@@ -13,7 +13,7 @@ import React, {FC, useEffect, useState} from 'react'
 import {useDispatch} from 'react-redux'
 import {getBuckets} from 'src/buckets/actions/thunks'
 import {event} from 'src/cloud/utils/reporting'
-import {keyboardCopyTriggered} from 'src/utils/keyboardCopy'
+import {keyboardCopyTriggered} from 'src/utils/crossPlatform'
 
 export const InstallDependencies: FC = () => {
   const dispatch = useDispatch()
@@ -62,37 +62,36 @@ sudo cp influxdb2-client-latest-linux-arm64/influx /usr/local/bin/
   )
 
   useEffect(() => {
+    const userSelection = () => {
+      return window.getSelection().toString()
+    }
+
+    const fireKeyboardCopyEvent = event => {
+      if (
+        keyboardCopyTriggered(event) &&
+        userSelection().includes('brew install')
+      ) {
+        logCopyCodeSnippetMac()
+      }
+      if (
+        keyboardCopyTriggered(event) &&
+        (userSelection().includes('Expand-Archive') ||
+          userSelection().includes('mv'))
+      ) {
+        logCopyCodeSnippetWindows()
+      }
+      if (
+        keyboardCopyTriggered(event) &&
+        (userSelection().includes('wget') ||
+          userSelection().includes('tar') ||
+          userSelection().includes('sudo'))
+      ) {
+        logCopyCodeSnippetLinux()
+      }
+    }
     document.addEventListener('keydown', fireKeyboardCopyEvent)
     return () => document.removeEventListener('keydown', fireKeyboardCopyEvent)
   }, [])
-
-  const userSelection = () => {
-    return window.getSelection().toString()
-  }
-
-  const fireKeyboardCopyEvent = event => {
-    if (
-      keyboardCopyTriggered(event) &&
-      userSelection().includes('brew install')
-    ) {
-      logCopyCodeSnippetMac()
-    }
-    if (
-      keyboardCopyTriggered(event) &&
-      (userSelection().includes('Expand-Archive') ||
-        userSelection().includes('mv'))
-    ) {
-      logCopyCodeSnippetWindows()
-    }
-    if (
-      keyboardCopyTriggered(event) &&
-      (userSelection().includes('wget') ||
-        userSelection().includes('tar') ||
-        userSelection().includes('sudo'))
-    ) {
-      logCopyCodeSnippetLinux()
-    }
-  }
 
   return (
     <>
