+++ src/utils/crossPlatform.ts
@@ -33,3 +33,32 @@ export const shouldOpenLinkInNewTab = (
 
   return false
 }
+
+export const keyboardCopyTriggered = (
+  event: KeyboardEvent | React.KeyboardEvent
+): boolean => {
+  let OS = OperatingSystems.Mac
+
+  if (window?.navigator?.userAgent.includes('Windows')) {
+    OS = OperatingSystems.Windows
+  }
+
+  if (window?.navigator?.userAgent.includes('Linux')) {
+    OS = OperatingSystems.Linux
+  }
+
+  // keyCode 67 is "C"
+  if (OS === OperatingSystems.Mac && event.metaKey && event.keyCode == 67) {
+    return true
+  }
+
+  if (
+    (OS === OperatingSystems.Windows || OS === OperatingSystems.Linux) &&
+    event.ctrlKey &&
+    event.keyCode == 67
+  ) {
+    return true
+  }
+
+  return false
+}
